mgmt-system/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.13
pydantic==1.10.7
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.10.4
psycopg2-binary==2.9.6
python-multipart==0.0.6
aiofiles==23.1.0
pytest==7.3.1
httpx==0.24.0
langchain==0.0.177
openai==0.27.6
tenacity==8.2.2 